% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_climate.R
\name{get_climate}
\alias{get_climate}
\title{get climate data according to lake name from GLTC dataset}
\usage{
get_climate(lake_name, types)
}
\arguments{
\item{lake_name}{a valid name of a lake in the GLTC dataset}

\item{types}{name for the climate data. All returned if missing.}
}
\value{
a lake data.frame, or an empty data.frame if no data exist
}
\description{
get climate data according to lake name from GLTC dataset
}
\examples{
get_climate_names()
get_climate('Victoria', types = c('Radiation.Shortwave.Summer', 'Air.Temp.Mean.Summer.NCEP'))
get_climate('Mendota')
get_climate('mendota','radiation.shortwave.summer')
}
\seealso{
\code{\link{get_climate_names}}, \code{\link{get_surface_temps}}
}

