% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_lake_names.R
\name{get_lake_names}
\alias{get_lake_names}
\title{get lake names in GLTC dataset}
\usage{
get_lake_names()
}
\value{
a character vector of valid lake names
}
\description{
get lake names in GLTC dataset
}
\examples{
get_lake_names()
}
\seealso{
\code{\link{get_climate_names}}, \code{\link{get_metadata_names}}
}

