% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_metadata_names.R
\name{get_metadata_names}
\alias{get_metadata_names}
\title{get metadata names in GLTC dataset}
\usage{
get_metadata_names(lake_name = NULL)
}
\arguments{
\item{lake_name}{a valid name of a lake in the GLTC dataset (see \code{\link{get_lake_names}}).}
}
\value{
a character vector of valid metadata variable names for all lakes.
If \code{lake_name} is specified, only the metadata fields populated for \code{lake_name}
are returned.
}
\description{
Lake metadata records were compiled for the Global Lake Temperature
Collaboration and can be accessed and returned as data.frames.
See associated publication and references therein for details.
}
\examples{
get_metadata_names()
get_metadata_names('Victoria')
}
\seealso{
\code{\link{get_metadata}}, \code{\link{get_climate_names}}, \code{\link{get_lake_names}},
}

