% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_surface_temps}
\alias{get_surface_temps}
\title{get lake data according to lake name}
\usage{
get_surface_temps(lake_name, type = "in_situ")
}
\arguments{
\item{lake_name}{a valid name of a lake in the GLTC database}

\item{type}{source for the data. Either "in_situ" or "satellite".}
}
\value{
a lake data.frame
}
\description{
get lake data according to lake name
}

