% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_climate}
\alias{get_climate}
\title{get climate data according to lake name from GLTC dataset}
\usage{
get_climate(lake_name, types = "all")
}
\arguments{
\item{lake_name}{a valid name of a lake in the GLTC dataset}

\item{types}{name for the climate data.}
}
\value{
a lake data.frame, or NULL if no data exist
}
\description{
get climate data according to lake name from GLTC dataset
}
\examples{
get_climate_names()
get_climate('Victoria', types = c('SWdown_Summer', 'NCEP 3 month'))
get_climate('Mendota')
}

