% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_climate_names.R
\name{get_climate_names}
\alias{get_climate_names}
\title{get climate data names in GLTC dataset}
\usage{
get_climate_names(lake_name = NULL)
}
\arguments{
\item{lake_name}{a valid name of a lake in the GLTC database}
}
\value{
a character vector of valid climate variable names
}
\description{
get climate data names in GLTC dataset
}
\examples{
get_climate_names()
}
\seealso{
\code{\link{get_climate}}, \code{\link{get_lake_names}}
}

