% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_climate_names}
\alias{get_climate_names}
\title{get climate data names in GLTC dataset}
\usage{
get_climate_names()
}
\value{
a character vector of valid climate variable names
}
\description{
get climate data names in GLTC dataset
}
\examples{
get_climate_names()
}
\seealso{
\code{\link{get_climate}}, \code{\link{get_lake_names}}
}

