% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_metadata}
\alias{get_metadata}
\title{get metadata for a lake in GLTC dataset}
\usage{
get_metadata(lake_name, metadata_name = NULL)
}
\arguments{
\item{lake_name}{a name of a lake in GLTC dataset}

\item{metadata_name}{a name of a metadata variable in GLTC dataset (optional)}
}
\value{
list with metadata (if metadata_name == NULL) or value if
metadata_name is specified
}
\description{
get metadata for a lake in GLTC dataset
}
\examples{
get_metadata('Victoria','Sampling depth')
get_metadata('Mendota')
}

