% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_metadata_names}
\alias{get_metadata_names}
\title{get metadata names in GLTC dataset}
\usage{
get_metadata_names(lake_name = NULL)
}
\arguments{
\item{lake_name}{}
}
\value{
a character vector of valid metadata variable names for all lakes.
If lake_name is specified, only the metadata fields populated for lake_name
are returned.
}
\description{
get metadata names in GLTC dataset
}
\examples{
get_metadata_names()
get_metadata_names('Victoria')
}

