% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_surface_temps}
\alias{get_surface_temps}
\title{get summer surface lake temperature data from GLTC dataset}
\usage{
get_surface_temps(lake_name, type)
}
\arguments{
\item{lake_name}{a valid name of a lake in the GLTC database}

\item{type}{source for the data. Either "In situ" or "Satellite".}
}
\value{
a lake data.frame, or NULL if no data exist
}
\description{
get summer surface lake temperature data from GLTC dataset
}
\examples{
get_surface_temps('Victoria','Satellite')
get_surface_temps('Mendota','In situ')
}

