% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_surface_temps.R
\name{get_surface_temps}
\alias{get_surface_temps}
\title{get summer surface lake temperature data from GLTC dataset}
\usage{
get_surface_temps(lake_name, type)
}
\arguments{
\item{lake_name}{a valid name of a lake in the GLTC database}

\item{type}{source for the data. Either "Lake.Temp.Summer.InSitu" or "Lake.Temp.Summer.Satellite".}
}
\value{
a lake data.frame (empty data.frame) if no data exist
}
\description{
get summer surface lake temperature data from GLTC dataset
}
\examples{
get_surface_temps('Victoria','Lake.Temp.Summer.Satellite')
get_surface_temps('Mendota','Lake.Temp.Summer.InSitu')
# retrieve from a lake site with multiple surface temperature sources:
get_surface_temps('Tahoe.Mid.Lake')
}
\seealso{
\code{\link{get_lake_names}}, \code{\link{get_climate}}
}

